chore: moved dev buttons to its own file
